LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/animation-trigger-shorthand.tentative-expected.txt


FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "once" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0% cover 100%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "cover 90%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "cover 10%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain 80%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain 20%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "once" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10% cover 90%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "once" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover 0%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "cover 10%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain 80%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain 20%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "once" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 20% contain 80% cover 10%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "once" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() once contain 0% contain 100% cover" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "alternate" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0% contain 100%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain 80%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain 20%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "alternate" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 20% contain 80%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "alternate" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 0%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain 10%" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "alternate" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain 10%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() alternate contain" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "alternate" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate contain" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "view() alternate" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "view()" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "alternate" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "view() alternate" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger'] = "repeat" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "repeat" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "repeat" should set animation-trigger-range-end assert_equals: animation-trigger-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "repeat" should set animation-trigger-range-start assert_equals: animation-trigger-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "repeat" should set animation-trigger-timeline assert_equals: animation-trigger-timeline should be canonical expected (string) "auto" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "repeat" should set animation-trigger-type assert_equals: animation-trigger-type should be canonical expected (string) "repeat" but got (undefined) undefined
FAIL e.style['animation-trigger'] = "repeat" should not set unrelated longhands assert_true: expected true got false
FAIL Property animation-trigger value 'view() once contain 0% contain 100% cover 0% cover 100%' assert_true: animation-trigger doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger value 'view() once contain 20% contain 80% cover 10% cover 90%' assert_true: animation-trigger doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger value 'view() once contain 20% contain 80% cover 0%' assert_true: animation-trigger doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger value 'view() once contain 20% contain 80% cover 10%' assert_true: animation-trigger doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger value 'view() once contain 0% contain 100%' assert_true: animation-trigger doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger value 'view() once' assert_true: animation-trigger doesn't seem to be supported in the computed style expected true got false
